[add]refs #36 coding-test7
